backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
